Remove explicit uses of `NSScreen`
