+++ Amethyst/Managers/FocusFollowsMouseManager.swift
@@ -17,6 +17,8 @@ protocol FocusFollowsMouseManagerDelegate: class {
 }
 
 final class FocusFollowsMouseManager<Delegate: FocusFollowsMouseManagerDelegate> {
+    typealias Screen = Delegate.Window.Screen
+
     weak var delegate: Delegate!
 
     private let userConfiguration: UserConfiguration
@@ -56,12 +58,12 @@ final class FocusFollowsMouseManager<Delegate: FocusFollowsMouseManagerDelegate>
             return
         }
 
-        guard let screen = NSScreen.screens.first(where: { $0.frame.contains(event.locationInWindow) }) else {
+        guard let screen = Screen.availableScreens.first(where: { $0.frame.contains(event.locationInWindow) }) else {
             return
         }
 
         var mousePoint = NSPointToCGPoint(event.locationInWindow)
-        mousePoint.y = NSScreen.globalHeight() - mousePoint.y + screen.frameIncludingDockAndMenu().origin.y
+        mousePoint.y = Screen.globalHeight() - mousePoint.y + screen.frameIncludingDockAndMenu().origin.y
 
         if let focusedWindow = Delegate.Window.currentlyFocused() {
             // If the point is already in the frame of the focused window do nothing.
