+++ Amethyst/Managers/WindowManager.swift
@@ -465,7 +465,7 @@ extension WindowManager: MouseStateKeeperDelegate {
 
         // need to flip mouse coordinate system to fit Amethyst https://stackoverflow.com/a/45289010/2063546
         let flippedPointerLocation = NSPointToCGPoint(NSEvent.mouseLocation)
-        let unflippedY = NSScreen.globalHeight() - flippedPointerLocation.y + screen.frameIncludingDockAndMenu().origin.y
+        let unflippedY = Screen.globalHeight() - flippedPointerLocation.y + screen.frameIncludingDockAndMenu().origin.y
         let pointerLocation = NSPointToCGPoint(NSPoint(x: flippedPointerLocation.x, y: unflippedY))
 
         if let screenManager: ScreenManager<Window> = focusedScreenManager(), let layout = screenManager.currentLayout {
